requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
pyinstaller>=6.4.0
pillow>=10.0.0
cryptography>=3.4.7

